Re-introduce the stack-reuse optimization
As predicted, with some extra thought it is possible to preserve the
stack-reuse optimization in lvm, the one having to do with CIST_FRESH.

First, this improves the performance by reducing stack usage for Lua-Lua
calls. Furthermore, it also fixes some bugs! For example, when we had
disabled the optimization, we had inadvertently broken vararg functions
(it had to do with the Protect messing with the L->top)