Remove the AOT_USE_TAILCALL option
In my experiments, the trampoline version proved to be faster than the
tail call version (most of the time -- a few benchmarks were faster
using tail calls). Since the trampoline version works even for older
compilers, and appears to be faster, I'm going to make it the default
and not bother with having an option for the tail call version.